Moved core interfaces to separate package
Refactored groundstation object
Introduced IPartOf interface and management

+++ exchange/dataaccess/src/main/java/org/hbird/exchange/dataaccess/TlePropagationRequest.java
@@ -140,8 +140,8 @@ public class TlePropagationRequest extends Command {
         return getArgumentValue(StandardArguments.DELTA_PROPAGATION, Long.class);
     }
 
-    public Double getStepSize() {
-        return getArgumentValue(StandardArguments.STEP_SIZE, Double.class);
+    public Long getStepSize() {
+        return getArgumentValue(StandardArguments.STEP_SIZE, Long.class);
     }
 
     public List<String> getLocations() {
